Add use of a user-specific configuration directory.
Files that are user-specific are all stored inside a new directory: ``usrcfg``

This directory is a sibling of ``usrhome`` : they must both be located inside
the same parent directory.

Storing the files in a separate directory will allow that directory to also
be a VCS repository, but a private one that users should not publish.