Properly deal with sub-shell
Identification of the USRHOME directory name variables was not being
initialized in zsh sub-shells because it does not source the ~/zprofile
then.  Therefore the user configuration initialization must be sourced
by ~/.zshrc, which is now done.

Note that it's probably possible to infer the value of all USRHOME directory
environment variable values without relying on a user configuration.  However,
at this early stage doing it this way will help detect situations where the
user configuration is not sourced when it should be and that will help render
the system a little more robust.  Later I might remove the need to store these
value inside the user configuration and the necessity for the setup script to
store the value of DIR_USRHOME.