Complete tracing support.  Sanitize PATH if required.
All Z Shell dot files now support zsh configuration file tracing.
That tracing is disabled by default and activated by the persistent
configuration (for the login shell) and can be dynamically activated
for the sub-shells.

The zshrc file now sanitizes the PATH.  It removes empty entries and
duplicates, warning if a modification was done.