Add usrhome-switch-path, support dynamic PATH manipulation,
Improve Homebrew support in x86 architecture by detecting the architecture
and only attempting to add the new directories when they are needed.

refactored: consider Homebrew and USRHOME support as something that is
installable at any time.  The usrhome_switch-path allows returning to the
original system PATH.  Later the user can install anything with the use-
commands, including Homebrew and USRHOME, allowing tests on various ordering
of the components in PATH.

The next step will be to provide the sanitization path as as sourceable command.