Starting the support for multiple types of shells
The ibin sourced files will have an extension when they are
specific to a specific shell, otherwise they will be expected to
be POSIX compliant.
The command aliases will pick the one that correspond to the
type of shell being used.
The goal being to support the Z Shell, as much as Bash, at least at first.
Later support for other shells could be added.
This will, also, document some of the differences between the features
of the various shells.