Refactored setup directory tree
The setup/template directory will grow into a tree of example files
for several things, therefore the directory is now split into a tree
that represents the location on the target, starting from user's home
directory.

The directories named 'dot' are special directories; they hold non-hidden
files that are supposed to be hidden and stored in the corresponding parent
directory.

The organization mimics the layout of usrhome itself.