CBR now supports a local cbr command
After executing `use-cbr`, the  `c`, `b` and `r` commands now check if there
is a cbr executable inside the local directory, and if so executes it, passing
the command code ('c', 'b', or 'r') to it.

That can be used for any purpose, one of them it to change the current
directory to the root of the depot and re-issue the command from there, like
this, written in the sub-directory under the project root:

  #!/bin/sh
  cd ..
  source "$USRHOME_DIR/ibin/do-cbr" $1

With this the c, b and r commands can be executed from the root and the
sub-directory.