Add use-emacs-for-man: use Emacs as man reader.
This supports 3 ways of using Emacs: as a client to Emacs daemon, as emacs
running in graphics mode inside another GUI window, or in terminal mode
running inside the current shell.

Emacs is a much better man reader than man.  This script activates using Emacs
for man.  The use-emacs-for-man script accepts an option to specify the
default mode of emacs for man.  Then implements a man() function that is
exposed and replaces the man command in the current shell.  That man() command
also supports options to specify which way to launch Emacs; if it is not
explicitly specified the function uses the default that was set by
use-emacs-for-man

This also provides the command 'ge' in the template.  It launches Emacs in GUI
mode from the current shell with its environment.  emacs for man uses it
unless the ge is found in the PATH.  Users can copy the template copy into
their PATH and modify it to fit their (possibly different) need.