Improved use-emacs-for-man: new startup options, better help
Now support the -G and -T options, that correspond to the -g and -t but start
Emacs quicker with less initialization with the -Q option.

Also support the _S option whcih does the same as -s except that it does not
delete a window; a better way to use man from within the shell of an
emacsclient frame.