Better option arg/error handling
Better argument handling for the injected man() function
and the use-emacs-for-man command.

The -h and --help options are now accepted and just print
the usage, they don't print the ERROR line.

The exit code for -h and --help is 1, but is 2 or 3 in
case of command line detected error.

The man() function also accepts the -h and --help option,
and only print the usage, but in the case of man(), it also
print an attention message, warning the user that man is not
the native man command and showing how to issue man from the
instrumented shell.