man -s and -S print error if topic not found and exit code is 4
The mechanism used to detect if the man topic was found is
probably not very robust: it checks into the *Message* buffer for the
string stating that the topic was formatted.  Looking for the buffer presence
does not seem to work; looks like the buffer is first created as a potential
output and when fails the buffer is still there.

Ideally Emacs man/woman logic would be updated to be able to report whether
the topic was found. And also probably updated to improve it a little.
But that's a different project.

I'll use the current logic for now, and see if it can be improved later.