+ zsh prompt model 3: zsh% or zsh# in bold green or red
Set USRHOME_PROMPT_MODEL to 3 to use a little more colorful prompt;
it shows the shell type, zsh, followed by % or # on the second line,
in bold green if the last command succeeded or bold red if it failed.

Showing the shell type help when capturing commands for logs: it explicitly
identifies the shell.