Added popping the tracing level.
I would have preferred finding a way to register the execution of a function
when a script ends.  I don't think there's a portable way to do that...

So the end of each script has to have the explicit call to pop the trace
level.

Maybe I can use some sort of exit trap.  Will investigate next.