Refactor to better support Bash & simplify a little
USRHOME_TRACE_SHELL_COMFIG is now defined by usrcfg/setfor-shell-tracing.sh
This is all that is required for USRHOME to activate (or not) the shell
configuration tracing.  The user configuration is somewhere else.

This file can then be sourced by ~/.zshenv --> usrhome/dot/zshenv.zsh
and Bash config file.

All alias and sinple functions (the cd, mdd) are now defined inside the
ibin/setfor-alias or the shell specific ones; ibin/setfor-zsh-alias or
ibin/setfor-bas-alias which source it.