Fix zsh prompt logic: honor user's choice.  Also change # 2 and #3
Prompt #2, which should probably become the default for zsh, has 2 lines.
The second line show zsh and % for normal user and # for root.
That line uses uses color to indicate the state of the last command and when
the shell is at root:

- Normal user: green when last command succeeds, bold red when it failed.
- root:        magenta when last command succeeds, bold red when it failed.

Prompt #3 is like #2 but it does not uses color one the second line.