usrhome-prompt-model-to improvements: Allow overriding prompt model
Modified the infrastructure logic that allows dynamically changing the prompt.
The code now uses a secondary variable that overrides the prompt model value,
allowing this to be used when a shell restart is required.

A shell restart is often not required, as it's the case for the current Bash
prompts and most zsh prompts supported by USRHOME.  Add ability to allow the
user configuration to request a zsh restart when changing the zsh prompt.

So this new value does not always needs to restart the zsh shell when changing
the zsh prompt, making it more convenient.