Reduce # of files. Fix sub-shell prompt model selection.
Eliminated the need for usrcfg/setfor-shell-tracing.sh by moving its value
setting logic inside usrcfg/setfor-all-config.sh and having
usrcfg/setfor-all-config.sh source the usrhome/ibin/shell-tracing.sh (some
form of a call-back).

Also modified logic of selecting the shell prompt model to allow the use of a
different prompt model number by zsh and bash while allowing the dynamic
selection of the prompt inside the shell.  This is done by checking if the
prompt override is on.  If it's not, then the prompt selection is done
according to the shell being used.

More of that logic still has to go inside the templates and I'll need to update
the diagram, the documentation and the code comments to ensure everything is consistent.