Better support macOS login Bash shells.
On macOS, it appears that the ~/.bash_profile is always executed, even when
the shell is not a login shell.  So on macOS, don't set USRHOME__IN_LOGIN to 1
inside usrhome/dot/bash_profile.bash and instead, provide 2 scripts that can
be used as launchers of "login bash shells" : these simulate a login Bash
shell on macOS from the point of view of USRHOME.  This allows testing scripts
meant to execute in a real login Bash.