Add mk-diffo, to support a stream of tests
The purpose of this tool is to setup a local file "database"
of source code and its related object file to monitor the
evolution of the assembler code of the file over the various changes
done to the source code of that file.

The tool relies on a set of environment variables that must first be set.
I will write an envfor-diffo script for setting these variables,
but for now they can be set with setenv.