Add au-format line filter to format audit log lines.
Implemented as a 3 stage filter that uses GNU AWK
and a file that holds the syscall lookup table.
That file is created (by an AWK script too) if it does not already exists.

Currently the file is created in the /tmp directory.
I will most likely change this to have the file reside in the USRHOME
directory tree and be ignored by Git.  To do later.