Simplify, using the au-format filter
The overall efficiency of this program is not great,
however it does not matter much since it is already filtering
the lines from the audit log, picking only the SYSERR and the AVC lines.

The au-format filter uses AWK and a pre-loaded lookup table to
translate the syscall numbers to their names.