Add clang-version utility
A shell script and a C++ program are supplied.

The C++ program is a single file C++ program; as such it can be built with GNU
make as long as it is stored inside a directory that has the same name as the
C++ file and the executable.

The C++ executable is *not placed in the PATH.
However the shell script has the same name as *is* placed on the PATH.
The purpose of the script is to check if the C++ program was built.
If the executable is already built, it uses it.
If the executable does not exist, the script builds it using GNU make and then
use it.

This way, the user does not even have to worry about compiling the C++
executable to use it; it is built whenever it is first used.

Fix commit mistake:

Removed the executable and commit the script instead.